+++ src/main/java/de/dm/intellij/liferay/language/freemarker/LiferayFtlVariableProvider.java
@@ -112,25 +112,9 @@ public class LiferayFtlVariableProvider extends FtlGlobalVariableProvider implem
                 }
 
                 //Provide Liferay Taglibs as predefined variables in their corresponding Freemarker namespaces
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-aui.tld", module, "liferay_aui"));
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-portlet-ext.tld", module, "liferay_portlet"));
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-portlet.tld", module, "portlet"));
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-security.tld", module, "liferay_security"));
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-theme.tld", module, "liferay_theme"));
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-ui.tld", module, "liferay_ui"));
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-util.tld", module, "liferay_util"));
-
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-product-navigation.tld", module, "liferay_product_navigation"));
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-journal.tld", module, "liferay_journal"));
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-flags.tld", module, "liferay_flags"));
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-layout.tld", module, "liferay_layout"));
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-site-navigation.tld", module, "liferay_site_navigation"));
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-asset.tld", module, "liferay_asset"));
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-map.tld", module, "liferay_map"));
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-item-selector.tld", module, "liferay_item_selector"));
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-expando.tld", module, "liferay_expando"));
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-frontend.tld", module, "liferay_frontend"));
-                result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/liferay-trash.tld", module, "liferay_trash"));
+                for (Map.Entry<String, String> taglibMapping : LiferayFreemarkerTaglibs.FTL_TAGLIB_MAPPINGS.entrySet()) {
+                    result.addAll(getTaglibSupportVariables("/com/liferay/tld/" + liferayVersionPrefix + "/" + taglibMapping.getKey(), module, taglibMapping.getValue()));
+                }
             }
 
             return result;
@@ -235,6 +219,10 @@ public class LiferayFtlVariableProvider extends FtlGlobalVariableProvider implem
 
     private List<? extends FtlVariable> getTaglibSupportVariables(@NotNull final String resource,  @NotNull final Module module, @NotNull @NonNls final String taglibPrefix) {
         URL url = TemplateMacroProcessorUtil.class.getResource(resource);
+        if (url == null) {
+            return Collections.emptyList();
+        }
+
         VirtualFile macroFile = VfsUtil.findFileByURL(url);
         XmlFile xmlFile = (XmlFile)PsiManager.getInstance(module.getProject()).findFile(macroFile);